+++ _strncmp.c
@@ -8,9 +8,9 @@
  *
  *Return: 0 if both equal, difference in their ASCII values on first mismatch
  */
-int _strncmp(char *s1, char *s2, unsigned int n)
+int _strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	for (i = 0; i < n; i++)
 		if (s1[i] != s2[i])
